MISC - found bug with force complete (fixed) - begin organizing game page into child components.
