+++ frontend/src/components/pages/Game/GameHeaderStatus.tsx
@@ -1,0 +1,9 @@
+const GameHeaderStatus = () => {
+  return (
+    <h1 className=" text-purple-300 text-[2em] text-opacity-75 text-center">
+      Game Status
+    </h1>
+  );
+};
+
+export default GameHeaderStatus;
